Renamed OverviewActiivty to MainActivity. Note that the mainactivity uses the activity_overview layout.
