+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/AppTheme">
         <activity
             android:screenOrientation="landscape"
-            android:name=".empaticaE4.OverviewActivity"
+            android:name=".empaticaE4.MainActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -33,7 +33,7 @@
             android:value="@integer/google_play_services_version" />
 
         <!--<activity-->
-            <!--android:name=".empaticaE4.OverviewActivity"-->
+            <!--android:name=".empaticaE4.MainActivity"-->
             <!--android:label="@string/app_name"-->
             <!--android:screenOrientation="landscape">-->
