+++ app/src/main/res/layout/activity_overview.xml
@@ -5,7 +5,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".empaticaE4.OverviewActivity">
+    tools:context=".empaticaE4.MainActivity">
 
     <ScrollView
         xmlns:android="http://schemas.android.com/apk/res/android"
